// adapters/mtproto/src/mtproto/gen/api.d.ts
export namespace api {
// ...
  export type Authorization<
    K extends keyof _Authorization = keyof _Authorization
  > = ToUnderscore<_Authorization, K>;
  export type _Authorization = {
    "authorization": {
      current?: true;                         // flags.0?true
      official_app?: true;                    // flags.1?true
      password_pending?: true;                // flags.2?true
      encrypted_requests_disabled?: true;     // flags.3?true
      call_requests_disabled?: true;          // flags.4?true
      unconfirmed?: true;                     // flags.5?true
      hash: bigint;                           // long
      device_model: string;                   // string
      platform: string;                       // string
      system_version: string;                 // string
      api_id: number;                         // int
      app_name: string;                       // string
      app_version: string;                    // string
      date_created: number;                   // int
      date_active: number;                    // int
      ip: string;                             // string
      country: string;                        // string
      region: string;                         // string
    },
  };
// ...
  export type InputCheckPasswordSRP<
    K extends keyof _InputCheckPasswordSRP = keyof _InputCheckPasswordSRP
  > = ToUnderscore<_InputCheckPasswordSRP, K>;
  export type _InputCheckPasswordSRP = {
    "inputCheckPasswordEmpty": {}
    "inputCheckPasswordSRP": {
      srp_id: bigint;                         // long
      A: Uint8Array;                          // bytes
      M1: Uint8Array;                         // bytes
    },
  };
// ...
  export type CodeSettings<
    K extends keyof _CodeSettings = keyof _CodeSettings
  > = ToUnderscore<_CodeSettings, K>;
  export type _CodeSettings = {
    "codeSettings": {
      allow_flashcall?: true;                 // flags.0?true
      current_number?: true;                  // flags.1?true
      allow_app_hash?: true;                  // flags.4?true
      allow_missed_call?: true;               // flags.5?true
      allow_firebase?: true;                  // flags.7?true
      logout_tokens?: Uint8Array[];           // flags.6?Vector<bytes>
      token?: string;                         // flags.8?string
      app_sandbox?: boolean;                  // flags.8?Bool
    },
  };
// ...
  export type EmailVerification<
    K extends keyof _EmailVerification = keyof _EmailVerification
  > = ToUnderscore<_EmailVerification, K>;
  export type _EmailVerification = {
    "emailVerificationCode": {
      code: string;                           // string
    },
    "emailVerificationGoogle": {
      token: string;                          // string
    },
    "emailVerificationApple": {
      token: string;                          // string
    },
  };
// ...
}
// ...
export namespace auth {
  export type SentCode<
    K extends keyof _SentCode = keyof _SentCode
  > = ToUnderscore<_SentCode, K>;
  export type _SentCode = {
    "auth.sentCode": {
      type: auth.SentCodeType;                // auth.SentCodeType
      phone_code_hash: string;                // string
      next_type?: auth.CodeType;              // flags.1?auth.CodeType
      timeout?: number;                       // flags.2?int
    },
    "auth.sentCodeSuccess": {
      authorization: auth.Authorization;      // auth.Authorization
    },
  };
// ...
  export type ExportedAuthorization<
    K extends keyof _ExportedAuthorization = keyof _ExportedAuthorization
  > = ToUnderscore<_ExportedAuthorization, K>;
  export type _ExportedAuthorization = {
    "auth.exportedAuthorization": {
      id: bigint;                             // long
      bytes: Uint8Array;                      // bytes
    },
  };

  export const exportedAuthorization: TLConstructor<_ExportedAuthorization, "auth.exportedAuthorization">;
  export type PasswordRecovery<
    K extends keyof _PasswordRecovery = keyof _PasswordRecovery
  > = ToUnderscore<_PasswordRecovery, K>;
  export type _PasswordRecovery = {
    "auth.passwordRecovery": {
      email_pattern: string;                  // string
    },
  };
// ...
  export type LoginToken<
    K extends keyof _LoginToken = keyof _LoginToken
  > = ToUnderscore<_LoginToken, K>;
  export type _LoginToken = {
    "auth.loginToken": {
      expires: number;                        // int
      token: Uint8Array;                      // bytes
    },
    "auth.loginTokenMigrateTo": {
      dc_id: number;                          // int
      token: Uint8Array;                      // bytes
    },
    "auth.loginTokenSuccess": {
      authorization: auth.Authorization;      // auth.Authorization
    },
  };

  export const loginToken: TLConstructor<_LoginToken, "auth.loginToken">;
  export const loginTokenMigrateTo: TLConstructor<_LoginToken, "auth.loginTokenMigrateTo">;
  export const loginTokenSuccess: TLConstructor<_LoginToken, "auth.loginTokenSuccess">;
  export type LoggedOut<
    K extends keyof _LoggedOut = keyof _LoggedOut
  > = ToUnderscore<_LoggedOut, K>;
  export type _LoggedOut = {
    "auth.loggedOut": {
      future_auth_token?: Uint8Array;         // flags.0?bytes
    },
  };

  export const loggedOut: TLConstructor<_LoggedOut, "auth.loggedOut">;
}
// ...
export namespace account {
// ...
  export type _Password = {
    "account.password": {
      has_recovery?: true;                    // flags.0?true
      has_secure_values?: true;               // flags.1?true
      has_password?: true;                    // flags.2?true
      current_algo?: api.PasswordKdfAlgo;     // flags.2?PasswordKdfAlgo
      srp_B?: Uint8Array;                     // flags.2?bytes
      srp_id?: bigint;                        // flags.2?long
      hint?: string;                          // flags.3?string
      email_unconfirmed_pattern?: string;     // flags.4?string
      new_algo: api.PasswordKdfAlgo;          // PasswordKdfAlgo
      new_secure_algo: api.SecurePasswordKdfAlgo; // SecurePasswordKdfAlgo
      secure_random: Uint8Array;              // bytes
      pending_reset_date?: number;            // flags.5?int
      login_email_pattern?: string;           // flags.6?string
    },
  };

  export const password: TLConstructor<_Password, "account.password">;
// ...
  export type PasswordInputSettings<
    K extends keyof _PasswordInputSettings = keyof _PasswordInputSettings
  > = ToUnderscore<_PasswordInputSettings, K>;
  export type _PasswordInputSettings = {
    "account.passwordInputSettings": {
      new_algo?: api.PasswordKdfAlgo;         // flags.0?PasswordKdfAlgo
      new_password_hash?: Uint8Array;         // flags.0?bytes
      hint?: string;                          // flags.0?string
      email?: string;                         // flags.1?string
      new_secure_settings?: api.SecureSecretSettings; // flags.2?SecureSecretSettings
    },
  };
// ...
}
// ...
export namespace auth {
  export const sendCode: TLApiMethod<"auth.sendCode", {
    phone_number: string                    // string
    api_id: number                          // int
    api_hash: string                        // string
    settings: api.CodeSettings              // CodeSettings
  }, SentCode>
  export const signUp: TLApiMethod<"auth.signUp", {
    no_joined_notifications?: true          // flags.0?true
    phone_number: string                    // string
    phone_code_hash: string                 // string
    first_name: string                      // string
    last_name: string                       // string
  }, Authorization>
  export const signIn: TLApiMethod<"auth.signIn", {
    phone_number: string                    // string
    phone_code_hash: string                 // string
    phone_code?: string                     // flags.0?string
    email_verification?: api.EmailVerification // flags.1?EmailVerification
  }, Authorization>
  export const logOut: TLApiMethod<"auth.logOut", void, LoggedOut>
  export const resetAuthorizations: TLApiMethod<"auth.resetAuthorizations", void, boolean>
  export const exportAuthorization: TLApiMethod<"auth.exportAuthorization", {
    dc_id: number                           // int
  }, ExportedAuthorization>
  export const importAuthorization: TLApiMethod<"auth.importAuthorization", {
    id: bigint                              // long
    bytes: Uint8Array                       // bytes
  }, Authorization>
  export const bindTempAuthKey: TLApiMethod<"auth.bindTempAuthKey", {
    perm_auth_key_id: bigint                // long
    nonce: bigint                           // long
    expires_at: number                      // int
    encrypted_message: Uint8Array           // bytes
  }, boolean>
  export const importBotAuthorization: TLApiMethod<"auth.importBotAuthorization", {
    flags: number                           // int
    api_id: number                          // int
    api_hash: string                        // string
    bot_auth_token: string                  // string
  }, Authorization>
  export const checkPassword: TLApiMethod<"auth.checkPassword", {
    password: api.InputCheckPasswordSRP     // InputCheckPasswordSRP
  }, Authorization>
  export const requestPasswordRecovery: TLApiMethod<"auth.requestPasswordRecovery", void, PasswordRecovery>
  export const recoverPassword: TLApiMethod<"auth.recoverPassword", {
    code: string                            // string
    new_settings?: account.PasswordInputSettings // flags.0?account.PasswordInputSettings
  }, Authorization>
  export const resendCode: TLApiMethod<"auth.resendCode", {
    phone_number: string                    // string
    phone_code_hash: string                 // string
  }, SentCode>
  export const cancelCode: TLApiMethod<"auth.cancelCode", {
    phone_number: string                    // string
    phone_code_hash: string                 // string
  }, boolean>
  export const dropTempAuthKeys: TLApiMethod<"auth.dropTempAuthKeys", {
    except_auth_keys: bigint[]              // Vector<long>
  }, boolean>
  export const exportLoginToken: TLApiMethod<"auth.exportLoginToken", {
    api_id: number                          // int
    api_hash: string                        // string
    except_ids: bigint[]                    // Vector<long>
  }, LoginToken>
  export const importLoginToken: TLApiMethod<"auth.importLoginToken", {
    token: Uint8Array                       // bytes
  }, LoginToken>
  export const acceptLoginToken: TLApiMethod<"auth.acceptLoginToken", {
    token: Uint8Array                       // bytes
  }, api.Authorization>
  export const checkRecoveryPassword: TLApiMethod<"auth.checkRecoveryPassword", {
    code: string                            // string
  }, boolean>
  export const importWebTokenAuthorization: TLApiMethod<"auth.importWebTokenAuthorization", {
    api_id: number                          // int
    api_hash: string                        // string
    web_auth_token: string                  // string
  }, Authorization>
  export const requestFirebaseSms: TLApiMethod<"auth.requestFirebaseSms", {
    phone_number: string                    // string
    phone_code_hash: string                 // string
    safety_net_token?: string               // flags.0?string
    ios_push_secret?: string                // flags.1?string
  }, boolean>
  export const resetLoginEmail: TLApiMethod<"auth.resetLoginEmail", {
    phone_number: string                    // string
    phone_code_hash: string                 // string
  }, SentCode>
}